test: consolidate css/html syntax unit tests into CssSyntax and HtmlSyntax
Merge walkCssTokens + walkCssTokensParser into CssSyntax.unittest.js and
walkHtmlTokens + buildHtmlAst into HtmlSyntax.unittest.js, mirroring the
lib/css/syntax.js and lib/html/syntax.js module names. Snapshot files are
renamed to match; html5lib.spectest.js is left separate.
