+++ test/CssSyntax.unittest.js
@@ -1,14 +1,37 @@
 "use strict";
 
+const fs = require("fs");
+const path = require("path");
 const {
 	Node,
 	NodeType,
 	SourceProcessor,
+	TT_AT_KEYWORD,
 	TT_BAD_STRING_TOKEN,
 	TT_BAD_URL_TOKEN,
+	TT_CDC,
+	TT_CDO,
+	TT_COLON,
+	TT_COMMA,
+	TT_COMMENT,
+	TT_DELIM,
+	TT_DIMENSION,
 	TT_EOF,
+	TT_FUNCTION,
+	TT_HASH,
+	TT_IDENTIFIER,
+	TT_LEFT_CURLY_BRACKET,
+	TT_LEFT_PARENTHESIS,
+	TT_LEFT_SQUARE_BRACKET,
+	TT_NUMBER,
+	TT_PERCENTAGE,
+	TT_RIGHT_CURLY_BRACKET,
+	TT_RIGHT_PARENTHESIS,
+	TT_RIGHT_SQUARE_BRACKET,
+	TT_SEMICOLON,
 	TT_STRING,
 	TT_URL,
+	TT_WHITESPACE,
 	Token,
 	TokenStream,
 	buildSkipSet,
@@ -22,6 +45,133 @@ const {
 	parseAStylesheetsContents,
 	walkTokens
 } = require("../lib/css/syntax");
+
+// Snapshot uses the spec-style kebab-case names for multi-word token types;
+// the tokenizer emits numeric `TT_*` values. Map between them so the existing
+// snapshot files stay valid.
+/** @type {Record<number, string>} */
+const TYPE_TO_PRINTED = {
+	[TT_WHITESPACE]: "whitespace",
+	[TT_COMMENT]: "comment",
+	[TT_URL]: "url",
+	[TT_LEFT_CURLY_BRACKET]: "left-curly-bracket",
+	[TT_RIGHT_CURLY_BRACKET]: "right-curly-bracket",
+	[TT_LEFT_PARENTHESIS]: "left-parenthesis",
+	[TT_RIGHT_PARENTHESIS]: "right-parenthesis",
+	[TT_LEFT_SQUARE_BRACKET]: "left-square-bracket",
+	[TT_RIGHT_SQUARE_BRACKET]: "right-square-bracket",
+	[TT_SEMICOLON]: "semicolon",
+	[TT_COMMA]: "comma",
+	[TT_AT_KEYWORD]: "at-keyword",
+	[TT_COLON]: "colon",
+	[TT_DELIM]: "delim",
+	[TT_NUMBER]: "number",
+	[TT_PERCENTAGE]: "percentage",
+	[TT_DIMENSION]: "dimension",
+	[TT_IDENTIFIER]: "identifier",
+	[TT_HASH]: "hash",
+	[TT_STRING]: "string",
+	[TT_FUNCTION]: "function",
+	[TT_CDO]: "cdo",
+	[TT_CDC]: "cdc",
+	[TT_BAD_STRING_TOKEN]: "bad-string-token",
+	[TT_BAD_URL_TOKEN]: "bad-url-token"
+};
+
+describe("walkTokens", () => {
+	const casesPath = path.resolve(__dirname, "./configCases/css/parsing/cases");
+	const tests = fs
+		.readdirSync(casesPath)
+		.filter((test) => /\.css/.test(test))
+		.map((item) => [
+			item,
+			fs.readFileSync(path.resolve(casesPath, item), "utf8")
+		]);
+
+	for (const [name, code] of tests) {
+		it(`should parse and print "${name}"`, () => {
+			const results = [];
+			// Drive the lexer core directly: a fresh `out` per call collects the
+			// raw token list (comments included); `walkTokens` returns undefined at EOF.
+			for (let pos = 0; ;) {
+				const t = walkTokens(
+					code,
+					pos,
+					/** @type {import("../lib/css/syntax").MutableToken} */ ({})
+				);
+				if (t === undefined) break;
+				pos = t.end;
+				const printed = TYPE_TO_PRINTED[t.type] || t.type;
+				if (t.type === TT_URL) {
+					results.push([
+						printed,
+						code.slice(t.start, t.end),
+						code.slice(t.contentStart, t.contentEnd)
+					]);
+				} else if (t.type === TT_HASH) {
+					results.push([printed, code.slice(t.start, t.end), t.isId]);
+				} else {
+					results.push([printed, code.slice(t.start, t.end)]);
+				}
+			}
+
+			expect(
+				results.filter((item) => item[0] !== "whitespace")
+			).toMatchSnapshot();
+			expect(results.map((item) => item[1]).join("")).toBe(code);
+		});
+	}
+});
+
+/**
+ * @param {string} input CSS source
+ * @returns {string} input reconstructed from token source slices
+ */
+const tokenRoundtrip = (input) => {
+	let out = "";
+	for (let pos = 0; ;) {
+		const t = walkTokens(
+			input,
+			pos,
+			/** @type {import("../lib/css/syntax").MutableToken} */ ({})
+		);
+		if (t === undefined) break;
+		pos = t.end;
+		out += input.slice(t.start, t.end);
+	}
+	return out;
+};
+
+// Regressions from the css-parsing-tests corpus: each input previously hung
+// the parser or dropped bytes from the token stream.
+describe("CssSyntax regressions", () => {
+	const NUL = String.fromCharCode(0);
+	const C1 = String.fromCharCode(0x80); // U+0080: an ident-start code point
+
+	it("does not hang on a literal U+0080 ident-start code point", () => {
+		expect(parseAListOfComponentValues(C1, 0, {})).toHaveLength(1);
+		expect(parseAListOfComponentValues(`a${C1}b`, 0, {})).toHaveLength(1);
+	});
+
+	it("does not hang on a backslash at EOF inside a url token", () => {
+		expect(parseAListOfComponentValues("url(a\\", 0, {})).toHaveLength(1);
+		expect(parseAListOfComponentValues("url(\\", 0, {})).toHaveLength(1);
+	});
+
+	it("emits an unterminated comment at EOF so token ranges cover all input", () => {
+		expect(tokenRoundtrip("a /* unterminated")).toBe("a /* unterminated");
+		expect(tokenRoundtrip("/* x")).toBe("/* x");
+	});
+
+	it("emits a string with a trailing backslash at EOF", () => {
+		expect(tokenRoundtrip('"ab\\')).toBe('"ab\\');
+		expect(tokenRoundtrip("url('a\\")).toBe("url('a\\");
+	});
+
+	it("never drops input bytes around a NUL code point", () => {
+		expect(tokenRoundtrip(`a${NUL}b`)).toBe(`a${NUL}b`);
+	});
+});
 
 /**
  * @param {string} src css source
@@ -41,7 +191,7 @@ const firstTokenType = (src) =>
 		)
 	).type;
 
-describe("walkCssTokens — component values (tokenToNode)", () => {
+describe("CssSyntax — component values (tokenToNode)", () => {
 	it("classifies each leaf token type", () => {
 		/**
 		 * @param {string} s source
@@ -167,7 +317,7 @@ describe("walkCssTokens — component values (tokenToNode)", () => {
 	});
 });
 
-describe("walkCssTokens — parser entry points", () => {
+describe("CssSyntax — parser entry points", () => {
 	it("parseADeclaration parses name, value and !important", () => {
 		const d = /** @type {import("../lib/css/syntax").Declaration} */ (
 			parseADeclaration("color: red")
@@ -274,7 +424,7 @@ describe("walkCssTokens — parser entry points", () => {
 	});
 });
 
-describe("walkCssTokens — Node / Token", () => {
+describe("CssSyntax — Node / Token", () => {
 	it("exposes range, loc and toString over the source", () => {
 		const decl = /** @type {import("../lib/css/syntax").Declaration} */ (
 			parseADeclaration("color: red")
@@ -314,7 +464,7 @@ describe("walkCssTokens — Node / Token", () => {
 	});
 });
 
-describe("walkCssTokens — SourceProcessor", () => {
+describe("CssSyntax — SourceProcessor", () => {
 	it("fires enter / exit visitors in source order", () => {
 		/** @type {string[]} */
 		const log = [];
@@ -438,7 +588,7 @@ describe("walkCssTokens — SourceProcessor", () => {
 	});
 });
 
-describe("walkCssTokens — nesting and error recovery", () => {
+describe("CssSyntax — nesting and error recovery", () => {
 	/**
 	 * @param {string} src css source
 	 * @returns {{ type: number, childRules: number, decls: number }[]} top-level rule summary
@@ -514,7 +664,7 @@ describe("walkCssTokens — nesting and error recovery", () => {
 	});
 });
 
-describe("walkCssTokens — tokenizer edge cases", () => {
+describe("CssSyntax — tokenizer edge cases", () => {
 	it("treats an unterminated string at EOF as a string token", () => {
 		expect(firstTokenType('"abc')).toBe(TT_STRING);
 	});
@@ -561,7 +711,7 @@ describe("walkCssTokens — tokenizer edge cases", () => {
 	});
 });
 
-describe("walkCssTokens — skip set (CssProcessOptions.skip)", () => {
+describe("CssSyntax — skip set (CssProcessOptions.skip)", () => {
 	// A bare declaration (parsed as block-contents, so there is no selector
 	// prelude to pollute the counts). Every leaf type appears both at the top
 	// level of the value and inside `bar(…)`, so one skip proves both the
@@ -807,7 +957,7 @@ describe("walkCssTokens — skip set (CssProcessOptions.skip)", () => {
 	});
 });
 
-describe("walkCssTokens — path accessors", () => {
+describe("CssSyntax — path accessors", () => {
 	/** @typedef {import("../lib/css/syntax").CssPath} CssPath */
 	const SRC =
 		"@media screen { .a { co\\6cor: red !important; background: url(x.png) var(--v, calc(1 + 2)); } } /* note */ .b { grid: [x] 1; }";
